Take home Test Part C
In this part of the Assignment, I had to create a mini calculator Unit Test using 2 Interfaces and 1 Abstract class to multiply and divide a certain number. If the number I used to test came back with the write result then the test was success, else a failure. Lastly, I had to check to see if the initialized field total value was initialized to 1. If it was then the test was a success, else a failure.